chore: Remove in-the-press section
