--- src/components/pages/homepage/press-logos.module.scss
@@ -1,19 +1,0 @@
-.logos {
-  align-items: center;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  &.extra-margin {
-    margin-bottom: spacer(32);
-  }
-  img {
-    filter: grayscale(1);
-    height: 1.5rem;
-    margin: 0 1.8rem 2rem 1.8rem; // ignore-style-rule
-    @media (min-width: $viewport-lg) {
-      height: 2rem;
-      margin: 0 spacer(32) spacer(64) spacer(32);
-      margin: 0 1.8rem 3.125rem 1.8rem; // ignore-style-rule
-    }
-  }
-}
