+++ src/components/pages/homepage/press.module.scss
@@ -5,6 +5,7 @@
 
 .header {
   @include type-size(300);
+  @include margin(64, bottom);
   @media (min-width: $viewport-lg) {
     @include type-size(400);
   }
@@ -12,4 +13,24 @@
 
 .section {
   @include margin(80, top);
+}
+
+.logos {
+  align-items: center;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  &.extra-margin {
+    margin-bottom: spacer(32);
+  }
+  img {
+    filter: grayscale(1);
+    height: 1.5rem;
+    margin: 0 1.8rem 2rem 1.8rem; // ignore-style-rule
+    @media (min-width: $viewport-lg) {
+      height: 2rem;
+      margin: 0 spacer(32) spacer(64) spacer(32);
+      margin: 0 1.8rem 3.125rem 1.8rem; // ignore-style-rule
+    }
+  }
 }
